requirements.txt
docx2pdf~=0.1.8
python-docx~=1.1.2
openpyxl~=3.1.5
qrcode~=8.0
pillow
opencv-python~=4.11.0.86
pyzbar~=0.1.9